% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_edges}
\alias{gene_edges}
\title{6 edges information between two genes of human}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{gene1}{human genes}
  \item{gene2}{human genes}
}
}
\source{
\url{https://thebiogrid.org/}
}
\usage{
gene_edges
}
\description{
A dataset containing the 6 interactions of genes
}
\keyword{datasets}
