% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node2vec-package.R
\docType{package}
\name{node2vec-package}
\alias{node2vec}
\alias{node2vec-package}
\title{node2vec: Algorithmic Framework for Representational Learning on Graphs}
\description{
Given any graph, the 'node2vec' algorithm can learn continuous feature representations for the nodes, which can then be used for various downstream machine learning tasks.The techniques are detailed in the paper "node2vec: Scalable Feature Learning for Networks" by Aditya Grover, Jure Leskovec(2016),available at <arXiv:1607.00653>.
}
\author{
\strong{Maintainer}: Yang Tian \email{tianyang1211@126.com}

Authors:
\itemize{
  \item Xu Li
  \item Jing Ren
}

}
\keyword{internal}
